Create Calculator
Prompts user to enter two numbers between 200 and 1000 and a symbol for operation, such as ‘+’, ‘-‘, ‘*’, and ‘/’ (use Scanner for input).  
Code uses a switch or nested if statements to perform the operation on the two numbers 
If the provided symbol is valid, displays the input data along with the result of the calculation to the console. Otherwise displays error message